Grisette.Lib.Synth.Context

Description

Documentation

type MonadContext (ctx :: Type -> Type) = (MonadError Text ctx, TryMerge ctx)

type MonadSymbolicContext (ctx :: Type -> Type) = (MonadContext ctx, MonadUnion ctx)

type MonadAngelicContext (ctx :: Type -> Type) = (MonadSymbolicContext ctx, MonadFresh ctx)

type ConcreteContext = Either Text

A concrete context is a context that does not do multi-path symbolic execution.

type SymbolicContext = ExceptT Text Union

A symbolic context is a context that does multi-path symbolic execution.

type AngelicContext = FreshT (ExceptT Text Union)

An angelic context is a context that does multi-path symbolic execution with angelic choices.