Grisette.Lib.Synth.Program.Concrete.Program

Description

Do not use plain concrete programs on operators that may generate multiple results that cannot be merged into a single result, or there can be significant performance issues.

Use the ProgMayMultiPath wrapper to allow multiple results. However, note that it is generally not a good idea to use non-simply-mergeable types as results, as it can lead to exponential blowup in the number of results.

Documentation

data Stmt op varId

Constructors

Stmt

Fields

stmtOp :: op

stmtArgIds :: [varId]

stmtResIds :: [varId]

data ProgArg varId ty

Constructors

ProgArg

Fields

progArgName :: Text

progArgId :: varId

progArgType :: ty

data ProgRes varId ty

Constructors

ProgRes

Fields

progResId :: varId

progResType :: ty

data Prog op varId ty

Constructors

Prog

Fields

progArgList :: [ProgArg varId ty]

progStmtList :: [Stmt op varId]

progResList :: [ProgRes varId ty]

data ProgPPrintError varId op

Constructors

StmtPPrintError (Stmt op varId) Int (OpPPrintError varId op)

ResultUndefined Int varId

ExtractSubProgError Text

prettyStmt :: (ConcreteVarId varId, OpPPrint op) => Int -> Stmt op varId -> StateT (VarIdMap varId) (Either (ProgPPrintError varId op)) (Doc ann)

prettyProg :: (ConcreteVarId varId, OpPPrint op, PPrint ty) => Text -> Prog op varId ty -> Either (ProgPPrintError varId op) (Doc ann)

stmtToDotNode :: (ConcreteVarId varId, OpPPrint op) => Text -> Int -> Stmt op varId -> StateT (VarIdToLabel varId) (Either (ProgPPrintError varId op)) (DotNode Text, [DotEdge Text])

progToDotSubGraph :: (ConcreteVarId varId, OpPPrint op, PPrint ty) => Text -> Prog op varId ty -> Either (ProgPPrintError varId op) (DotSubGraph Text)

class ProgPPrint prog where

Methods

pformatProg :: Text -> prog -> Either (Doc ann) (Doc ann)

class ProgToDot prog where

Methods

toDotProg :: Text -> prog -> DotSubGraph Text

eliminateDeadCode :: ConcreteVarId varId => Prog op varId ty -> Prog op varId ty

eliminateProgTableDeadCode :: ConcreteVarId varId => SymbolTable (Prog op varId ty) -> SymbolTable (Prog op varId ty)