Grisette.Lib.Synth.Reasoning.Parallel.Scheduler.Scheduler

Documentation

data Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher where

Constructors

Scheduler

Fields

:: forall sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher. { config :: SchedulerConfig sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher

, randGen :: AtomicGenM StdGen

, nodeInfo :: IORef (HashMap NodeId (NodeInfo sketchSpec))

, nodeStates :: IORef (HashMap NodeId (NodeState conProg symSemObj symVal conSemObj conVal matcher))

, nodeQueue :: IORef BiasedQueue

, nodeSplitQueue :: IORef BiasedQueue

, processes :: IORef (HashMap Int32 Process)

, processToNode :: IORef (HashMap Int32 NodeId)

, nodeToProcess :: IORef (HashMap NodeId Int32)

, dcTree :: IORef (DCTree (SymbolTable sketchSpec))

, currentMinimalCost :: IORef (Maybe Int)

, queueLock :: MVar ()

, stopped :: IORef Bool

, schedulerStartTime :: UTCTime

, nodeStatusSets :: IORef (HashMap Int (HashMap StatusType (HashSet NodeId)))

, depthToNodes :: IORef (HashMap Int (HashSet NodeId))

, splitNodesByDepth :: IORef (HashMap Int (HashSet NodeId))

, firstNotFullySplitDepth :: IORef Int

, everStartedNodesByDepth :: IORef (HashMap Int (HashSet NodeId))

} -> Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher

data NodeInfo sketchSpec

Constructors

NodeInfo

Fields

nodeSplitted :: Bool

nodeSketchTable :: SymbolTable sketchSpec

nodePriority :: Priority

nodeTimeoutSeconds :: Int

newScheduler :: SchedulerConfig sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO (Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher)

getCPid :: HasCallStack => Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO Int32

getProcessByCPid :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> Int32 -> IO Process

getProcess :: HasCallStack => Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO Process

getCurrentMinimalCost :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO (Maybe Int)

getCurrentTimeout :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO Int

getCurrentElapsedTime :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO NominalDiffTime

getStatus :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO (NodeStatus conProg)

getNodeInfo :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO (NodeInfo sketchSpec)

getDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO Int

getIsSplitted :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO Bool

getSketchTable :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO (SymbolTable sketchSpec)

getPriority :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO Priority

getNumRunningProcess :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO Int

getNumQueuedProcess :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO Int

setIsSplitted :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> Bool -> IO ()

setPriority :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> Priority -> IO ()

setTimeout :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> Int -> IO ()

removeProcessByCPid :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> Int32 -> IO ()

removeProcess :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> IO ()

updateCurrentMinimalCost :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> Maybe Int -> IO ()

getNodesByStatus :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> StatusType -> IO (HashSet NodeId)

getNodesByStatusAndDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> StatusType -> Int -> IO (HashSet NodeId)

getNodesByDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> Int -> IO (HashSet NodeId)

updateNodeState :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> NodeState conProg symSemObj symVal conSemObj conVal matcher -> IO ()

getSplitNodesByDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> Int -> IO (HashSet NodeId)

Get all nodes that have been split at a specific depth

isSplitAtDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> NodeId -> Int -> IO Bool

Check if a node has been split at a specific depth

updateFirstNotFullySplitDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO ()

Update the first not fully split depth by checking if the current depth is fully split, and if so, advancing to the next depth

getFirstNotFullySplitDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO Int

Get the first depth that is not fully split

resetFirstNotFullySplitDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO ()

Reset the first not fully split depth to 0 (usually when adding a new root node)

isDepthFullySplit :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> Int -> IO Bool

Check if a specific depth is fully split

getEverStartedNodes :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> IO (HashSet NodeId)

Get all nodes that have ever been started across all depths

getEverStartedNodesByDepth :: Scheduler sketchSpec sketch conProg costObj cost symSemObj symVal conSemObj conVal matcher -> Int -> IO (HashSet NodeId)

Get all nodes that have ever been started at a specific depth